requirements.txt
faiss-cpu
numpy
scikit-learn
sentence-transformers
transformers
torch
flask
matplotlib
pandas
tqdm
python-dotenv
docx
datasets
seaborn